requirements.txt
numpy==1.17.3
pandas==1.1.0
wxPython==4.0.4
PyPDF2==1.26.0
pdfminer.six==20181108
tabula-py==2.1.1
pdf2image==1.10.0
pdfplumber==0.5.14
XlsxWriter==1.2.8
pywin32==300